ReplayStorage Implementations + Tests

+++ retrayed-api/src/main/java/dev/etrayed/retrayed/api/Replay.java
@@ -8,7 +8,7 @@ import org.jetbrains.annotations.NotNull;
  */
 public interface Replay {
 
-    long id();
+    int id();
 
     @NotNull
     BlockingEventIterator eventIterator();
